rendering/testing on commit using Travis CI
All commits to the OpenBEL/language repository will trigger a Travis CI
build that will render BEL specifications to HTML. The exit code of the
tools/scripts/build-html.sh script determines the status of the build.

The HTML result will not be published, but it will be output in the
logs.